+++ CHANGELOG.md
@@ -6,9 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-## [0.3.0] - 2018-10-23
+## [0.3.0/0.3.3] - 2018-10-23
 ### Added
 - Redux-saga synchronization methods
+### Changed 
+- Fixed return for syncObjects
+- Saga name on saga actions types
+- Fixed config iteration on Check_Sync Saga
 
 ## [0.2.0/0.2.10] - 2018-10-22
 ### Added